Main vs all aggregations (#225)
* parsing notes to determine main set
* add in main attr to card WIP
* add in_main, total size; update tests
* fix app mappers, readme
* consistency; add count and update_sizes methods
* refactor: update return types to i64 for
* csv notes on boosterTypes, promoTypes
* booster promo types notes
* fix main set assignment, size calcs
* in main use with sort number
* fix card entity in main req
* set booster counts notes,
* is canon promo type check
* allow univ beyond, ff promos
* update scry version
* Update scry/src/card/mapper.rs
* Update scry/src/price/repository.rs

+++ scry/src/database.rs
@@ -26,9 +26,9 @@ impl ConnectionPool {
     pub async fn execute_query_builder(
         &self,
         mut builder: QueryBuilder<'_, sqlx::Postgres>,
-    ) -> Result<u64> {
+    ) -> Result<i64> {
         let result = builder.build().execute(&*self.pool).await?;
-        Ok(result.rows_affected())
+        Ok(result.rows_affected() as i64)
     }
 
     pub async fn count(&self, query: &str) -> Result<i64> {
